remove sapling.rs migrate some functionality into traits.rs
