+++ lib/src/wallet/keys.rs
@@ -14,7 +14,6 @@ use zingoconfig::ZingoConfig;
 
 pub(crate) mod extended_transparent;
 pub(crate) mod orchard;
-pub(crate) mod sapling;
 pub(crate) mod unified;
 
 /// Sha256(Sha256(value))
